moved FakeRawMemcacheClient to ".test" package in prod module
to avoid duplication (no more two copies of this class).
this could be a temporary solution until "folsom-test" module is
introduced.
